requirements.txt
hypercorn
Quart
Flask-Script
Flask-Failsafe
python-dotenv
aiohttp
aioredis
aio-pika
cryptography
pytz<2021.0
tortoise-orm
python-dateutil
websockets
asyncmy
